youtube-transciber/requirement.txt
youtube_transcript_api
streamlit
google-generativeai
python-dotenv
langchain
pathlib
